+++ src/main/java/com/likelionking/sbbstudy/domain/comment/repository/CommentRepository.java
@@ -1,7 +1,6 @@
 package com.likelionking.sbbstudy.domain.comment.repository;
 
-import com.likelionking.sbbstudy.domain.article.domain.Article;
-import com.likelionking.sbbstudy.domain.domain.Comment;
+import com.likelionking.sbbstudy.domain.comment.domain.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CommentRepository extends JpaRepository<Comment, Long> {
