+++ src/main/java/com/likelionking/sbbstudy/domain/comment/service/CommentService.java
@@ -2,8 +2,8 @@ package com.likelionking.sbbstudy.domain.comment.service;
 
 import com.likelionking.sbbstudy.domain.article.domain.Article;
 import com.likelionking.sbbstudy.domain.comment.repository.CommentRepository;
-import com.likelionking.sbbstudy.domain.domain.Comment;
-import com.likelionking.sbbstudy.domain.domain.CommentForm;
+import com.likelionking.sbbstudy.domain.comment.domain.Comment;
+import com.likelionking.sbbstudy.domain.comment.domain.CommentForm;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
